Add findShorty interactor
New interactor resposible to to interact with the shorties
repository and retrieve shorties by their shortcode.

Returns shorty when it's found or `ShortyNotFoundError` when
there is no shorty in the database with the requested `shortcode`.

Also, and because we want to count all the access to each shorty,
it uses a specific method in the repository that increments the
shorty `access_count` attribute.