Improve shortcode generation logic
In the beginning I used Faker to easily generate random strings
that respected the requested regex, but it's not really an approach
that should be used in production environments.

A better approach is to use `SecureRandom` and generate the shortcode
from the `urlsafe_base64` method. It generates a 6 characters string
but it can contain the `-` character which is not allowed. To fix that
simply added a call to `tr` to replace all `-` with `_`.